Reclaim

See merge request sdis1/2021/t4/g04!5

+++ proj1/src/peer/PeerInit.java
@@ -10,7 +10,7 @@ public interface PeerInit extends Remote {
 
     void delete(String filepath) throws RemoteException;
 
-    void reclaim(int diskspace) throws RemoteException;
+    void reclaim(double diskspace) throws RemoteException;
 
     String state() throws RemoteException;
 }
